### 2021-04-28 #3213 @Kevin-Mattheus-Moerman
@mark-koren since the error message relates to the ORCID profiles, can you please double check these and remove any blank entries for them.
