### 2021-05-18 #3213 @mark-koren
Ok cool, thanks for the update! A labmate published a related (Julia) package through Joss and the reviewers were
Jarvist Moore Frost (jarvist) and Sambit Kumar Dash (sambitdash), so maybe they could be good options, if they also do python.
