+++ src/ast_toolbox/simulators/crazy_trolley/crazy_trolley.py
@@ -1,0 +1,845 @@
+import dill
+import numpy as np
+from matplotlib import pyplot as plt
+import pdb
+import time
+
+def debug_print(var_list, global_scope, local_scope):
+    print_str = ''
+    for var in var_list:
+        print_str += (var + ': ' + repr(eval(var, global_scope, local_scope)) + ', ')
+
+    # print(print_str[:-2])
+
+class TrackList(object):
+    # Linked list of track or intersection cells
+
+    def __init__(self):
+        self._first = None
+        self._last = None
+        self.intersections = []
+
+    @property
+    def first(self):
+        return self._first
+
+    @first.setter
+    def first(self, value):
+        self._first = value
+        if self._last is None:
+            self._last = value
+
+    @property
+    def last(self):
+        return self._last
+
+    @last.setter
+    def last(self, value):
+        # Append to linked list, shift end of list pointer
+        self._last.next = value
+        while self._last.next is not None:
+            self._last = self._last.next
+
+    @property
+    def num_intersections(self):
+        return len(self.intersections)
+
+class TrackCell(object):
+
+    def __init__(self, x, y, p_obstruction = 0.0, p_reward=0.0):
+        self._x = x
+        self._y = y
+        self.obstruction = (np.random.rand() <= p_obstruction)
+        self.reward = (not self.obstruction and (np.random.rand() <= p_reward))
+        self._next = None
+
+    @property
+    def x(self):
+        return self._x
+
+    @x.setter
+    def x(self, value):
+        self._x = value
+
+    @property
+    def y(self):
+        return self._y
+
+    @y.setter
+    def y(self, value):
+        self._y = value
+
+    @property
+    def next(self):
+        return self._next
+
+    @next.setter
+    def next(self, value):
+        self._next = value
+
+class IntersectionCell(TrackCell):
+
+    def __init__(self, x, y, next_track_num=0):
+        self._next_track_num = next_track_num
+        self._next_track_starts = []
+        super().__init__(x,y, p_obstruction=0.0, p_reward=0.0)
+
+    @property
+    def next_track_num(self):
+        return self._next_track_num
+
+    @next_track_num.setter
+    def next_track_num(self, value):
+        self._next_track_num = value
+
+    @property
+    def next(self):
+        return self._next[self._next_track_num]
+
+    @next.setter
+    def next(self, value):
+        self._next = value
+
+    @property
+    def num_tracks(self):
+        return len(self._next)
+
+    @property
+    def next_track_start(self):
+        return self._next_track_starts[self._next_track_num]
+
+    @property
+    def next_track_list(self):
+        return self._next
+
+    @property
+    def next_track_starts(self):
+        return self._next_track_starts
+
+def print_frame(frame):
+    # Helper function to print a frame in one line, with standardized settings
+    plt.imshow(frame)
+    plt.show()
+
+class CrazyTrolleyHeadlessGame:
+    def __init__(self, height=16, width=32, hardest_level=50, init_p_split=0.3):
+
+        self.level_bonus = 100
+        self.gem_bonus = 50
+        self.initial_lives = 1
+
+        # self.background = None
+        # reset game tracking variables
+        self._update_counter = 0
+        self._level = 0
+        self._cell = None
+        self._lives = self.initial_lives
+        self._score = 0
+        self.intersection_num = 0
+        self.intersection_setting = 0
+
+        self._height = height
+        self._width = width
+
+        sigmoid = lambda x: 1 / (1 + np.exp(-x))
+        # Return 0 if on_clear_path
+        self.p_obstruction = lambda level, on_clear_path=False: 0 if on_clear_path else sigmoid(-5 + level / 2) / 2
[…]
