### 2021-05-19 dbf64d5
New settings to train DQN on crazy trolley
