### 2021-05-20 40b9c63
New training setting
