+++ src/ast_toolbox/simulators/crazy_trolley/train_crazy_trolley.py
@@ -38,7 +38,7 @@ def run_task(snapshot_config, variant_data, *_):
 
     """
     with LocalTFRunner(snapshot_config=snapshot_config) as runner:
-        n_epochs = 1
+        n_epochs = 100
         n_epoch_cycles = 20
         sampler_batch_size = 500
         num_timesteps = n_epochs * n_epoch_cycles * sampler_batch_size
@@ -125,7 +125,7 @@ def _args(buffer_size):
 replay_buffer_size = _args.main(standalone_mode=False)
 run_experiment(
     run_task,
-    n_parallel=16,
+    n_parallel=32,
     snapshot_mode='last',
     seed=1,
     plot=False,
