### 2021-05-21 ab93296
Major penalty on dying
