### 2021-05-21 7daab83
Major penalty on dying
