### 2021-05-24 a8f5e51
Update colors and path length
