### 2021-05-25 #3312 @sjpfenninger
@mark-koren Can you ensure that the reference list includes DOIs or URLs wherever possible?
