### 2021-05-25 3f5f479
Try decaying learning rate
