+++ joss/paper.bib
@@ -60,6 +60,7 @@
   author={Lee, Ritchie and Kochenderfer, Mykel J. and Mengshoel, Ole J. and Brat, Guillaume P. and Owen, Michael P.},
   booktitle=dasc,
   year={2015},
+  doi={10.1109/DASC.2015.7311613},
 }
 
 @article{lee2020adaptive,
@@ -68,7 +69,8 @@
   journal={Journal of Artificial Intelligence Research},
   volume={69},
   pages={1165--1201},
-  year={2020}
+  year={2020},
+  doi={10.1613/jair.1.12190},
 }
 
 @Book{DMU,
