+++ joss/paper.bib
@@ -60,6 +60,7 @@
   author={Lee, Ritchie and Kochenderfer, Mykel J. and Mengshoel, Ole J. and Brat, Guillaume P. and Owen, Michael P.},
   booktitle=dasc,
   year={2015},
+  doi={10.1109/DASC.2015.7311613},
 }
 
 @article{lee2020adaptive,
@@ -69,6 +70,7 @@
   volume={69},
   pages={1165--1201},
   year={2020},
+  doi={10.1613/jair.1.12190},
 }
 
 @Book{DMU,
@@ -77,14 +79,16 @@
     Publisher                = {{MIT} Press},
   	Year                     = {2015},
   	Chapter                  = {Model Uncertainty},
-  	Pages                    = {113-132}
+  	Pages                    = {113-132},
+  	doi = {10.7551/mitpress/10187.001.0001},
 }
 
 @article{koren2020formulation,
   title={The Adaptive Stress Testing Formulation},
   author={Koren, Mark and Corso, Anthony and Kochenderfer, Mykel J.},
   journal={arXiv preprint arXiv:2004.04293},
-  year={2020}
+  year={2020},
+  doi = {},
 }
 
 @article{brockman2016openai,
@@ -130,6 +134,7 @@
     author={Mark Koren and Mykel J. Kochenderfer},
     journal={arXiv preprint arXiv:2004.04292},
     year={2020},
+    doi = {10.1109/ITSC45102.2020.9294729},
 }
 
 @inproceedings{koren2018adaptive,
@@ -137,7 +142,8 @@
   author={Koren, Mark and Alsaif, Saud and Lee, Ritchie and Kochenderfer, Mykel J.},
   booktitle={IEEE Intelligent Vehicles Symposium},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/IVS.2018.8500400},
 }
 
 @inproceedings{corso2019adaptive,
@@ -146,14 +152,16 @@
   booktitle={IEEE Intelligent Transportation Systems Conference (ITSC)},
   pages={163--168},
   year={2019},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/ITSC.2019.8917242},
 }
 
 @article{corso2020interpretable,
   title={Interpretable Safety Validation for Autonomous Vehicles},
   author={Corso, Anthony and Kochenderfer, Mykel J},
   journal={arXiv preprint arXiv:2004.06805},
-  year={2020}
+  year={2020},
+  doi = {10.1109/ITSC45102.2020.9294490},
 }
 
 @inproceedings{koren2019efficient,
@@ -162,7 +170,8 @@
   booktitle={IEEE Intelligent Transportation Systems Conference (ITSC)},
   pages={4178--4183},
   year={2019},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/ITSC.2019.8917403},
 }
 
 @inproceedings{moss2020adaptive,
@@ -170,11 +179,13 @@
   author={Moss, Robert J and Lee, Ritchie and Visser, Nicholas and Hochwarth, Joachim and Lopez, James G and Kochenderfer, Mykel J},
   booktitle=dasc,
   year={2020},
+  doi = {10.1109/DASC50938.2020.9256730},
 }
 
 @article{julian2020validation,
   title={Validation of Image-Based Neural Network Controllers through Adaptive Stress Testing},
   author={Julian, Kyle D and Lee, Ritchie and Kochenderfer, Mykel J},
   journal={arXiv preprint arXiv:2003.02381},
-  year={2020}
+  year={2020},
+  doi = {10.1109/itsc45102.2020.9294549},
 }
