### 2021-05-28 957ced3
Get crazy trolley ready for validation with MCTS
