+++ src/ast_toolbox/simulators/crazy_trolley/crazy_trolley_ast.py
@@ -1,0 +1,304 @@
+"""Class template to wrap a simulator for interaction with AST."""
+import numpy as np
+import tensorflow as tf
+import warnings
+
+from collections import deque
+from skimage import color, img_as_ubyte
+from skimage.transform import resize
+
+from garage.experiment import Snapshotter
+
+from ast_toolbox.simulators import ASTSimulator
+from ast_toolbox.simulators.crazy_trolley.crazy_trolley import CrazyTrolleyRenderedGame
+
+
+class CrazyTrolleyASTSimulator(ASTSimulator):
+    """Class template to wrap a simulator for interaction with AST.
+
+    This class already tracks the simulator options to return the correct observation type. In addition,
+    `max_path_length` and `self._path_length` are handled by this parent class.
+
+    Parameters
+    ----------
+    blackbox_sim_state : bool, optional
+        True if the true simulation state can not be observed, in which case actions and the initial conditions are
+        used as the observation. False if the simulation state can be observed, in which case it will be used.
+    open_loop : bool, optional
+        True if the simulation is open-loop, meaning that AST must generate all actions ahead of time, instead
+        of being able to output an action in sync with the simulator, getting an observation back before
+        the next action is generated. False to get interactive control, which requires that `blackbox_sim_state`
+        is also False.
+    fixed_init_state : bool, optional
+        True if the initial state is fixed, False to sample the initial state for each rollout from the observaation
+        space.
+    max_path_length : int, optional
+        Maximum length of a single rollout.
+    """
+
+    def __init__(self, height=16, width=32, from_pixels=True, rgb=True, random_level=False,
+                 skip=5, noop=30, stack_frames=5, max_and_skip=True, grayscale=True, resize = None,
+                 **kwargs):
+
+        super().__init__(**kwargs)
+
+        snapshotter = Snapshotter()
+        self.session = tf.compat.v1.Session()
+        with self.session.as_default():  # optional, only for TensorFlow
+            data = snapshotter.load('./examples/crazy_trolley/crazy_trolley_suts/experiment_2021_05_24_11_26_48_0001')
+            self.agent = data['algo'].policy
+
+            self.skip = skip
+            self.noop = noop
+            self.max_and_skip = max_and_skip
+            self.stack_frames = stack_frames
+            self.player_observation = deque(maxlen=self.stack_frames)
+            self.grayscale = grayscale
+            self.resize = resize
+            if self.resize is None:
+                self.resize = (height, width)
+            self._obs_buffer = np.zeros((2,) + self.resize,
+                                        dtype=np.uint8)
+
+            self.renderer = CrazyTrolleyRenderedGame(ax=None, height=height, width=width, rgb=rgb)
+            self.game = self.renderer.game
+
+            self.score = 0
+
+            self.rendered = False
+            self.random_level = random_level
+
+
+    def simulate(self, actions, s_0):
+        """Run a full simulation given the AST solver's actions and initial conditions.
+
+        `simulate` takes in the AST solver's actions and the initial conditions. It should return two values: a
+        terminal index and an array of relevant simulation information.
+
+        Parameters
+        ----------
+        actions : list[array_like]
+            A sequential list of actions taken by the AST Solver which deterministically control the simulation.
+        s_0 : array_like
+            An array specifying the initial conditions to set the simulator to.
+
+        Returns
+        -------
+        terminal_index : int
+            The index of the action that resulted in a state in the goal set E. If no state is found
+            terminal_index should be returned as -1.
+        array_like
+            An array of relevant simulator info, which can then be used for analysis or diagnostics.
+
+        """
+        self.reset(s_0)
+        while self._path_length <= self.c_max_path_length:
+            self._path_length += 1
+            action = actions[self._path_length]
+            self.closed_loop_step(action)
+
+            if self.game.game_over():
[…]
