### 2021-05-28 b03c814
Get crazy trolley ready for validation with MCTS
