### 2021-06-01 6592069
Remove stack frames
