### 2021-06-01 148588d
Remove stack frames
