### 2021-06-01 #3312 @giodegas
@mark-koren I tried to installi it, but got errors:

    pip3 install ast-toolbox

ends with:

```
  Could not find a version that satisfies the requirement tensorflow<1.16,>=1.15.0 (from garage==2019.10.1->ast-toolbox) (from versions: 1.13.0rc1, 1.13.0rc2, 1.13.1, 1.13.2, 1.14.0rc0, 1.14.0rc1, 1.14.0, 2.0.0a0, 2.0.0b0, 2.0.0b1)
No matching distribution found for tensorflow<1.16,>=1.15.0 (from garage==2019.10.1->ast-toolbox)
```

    pip3 install https://github.com/sisl/AdaptiveStressTestingToolbox.git

ends with:

`  Cannot unpack file /tmp/pip-unpack-79upke_s/AdaptiveStressTestingToolbox.git (downloaded from /tmp/pip-req-build-9pkfxc7f, content-type: text/html; charset=utf-8); cannot detect archive format
Cannot determine archive format of /tmp/pip-req-build-9pkfxc7f`

My Python3:
```
Python 3.7.3 (default, Jan 22 2021, 20:04:44)
[GCC 8.3.0] on linux
```
