### 2021-06-02 6093ee7
Add custom grayscale and resize wrapper envs using PIL because scikit implementation was broken
