### 2021-06-02 769ca87
New colors optimized for max difference in grayscale space
