### 2021-06-02 b77f3fe
Use new grayscale and resize custome environments
