### 2021-06-02 a536161
Update installation from git repo instructions
