+++ setup.py
@@ -95,7 +95,7 @@ setup(
     keywords=[
         # eg: 'keyword1', 'keyword2', 'keyword3',
     ],
-    python_requires='~=3.6',
+    python_requires='>=3.6, !=3.7.*, !=3.8.*, !=3.9.*, !=3.10.*',
     # python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*',
     install_requires=REQUIRED,
     extras_require=EXTRAS,
