+++ .bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 2020.09.01.6
+current_version = 2020.09.01.7
 parse = (?P<year>\d+)\.(?P<month>\d+)\.(?P<day>\d+)\.(?:(?P<release>[a-z]+))?(?P<relnumber>\d+)?
 serialize = 
 	{year}.{month}.{day}.{release}{relnumber}
