### 2021-06-02 #3312 @mark-koren
@giodegas Try using python 3.6*. If that is an issue, let me know -- I was able to successfully install the toolbox using the latest version of 3.7 (3.7.10), though using 3.7.3 I also get the error you did. I have updated the toolbox for now to enforce using 3.6.

*This is an inherited dependency from garage, but they have since released an update that will allow us to support higher versions of python. The next planned update for this toolbox will be to use the newer version of garage and support more python versions, but it is a lot of work and will be sometime this summer.
