+++ src/ast_toolbox/simulators/crazy_trolley/train_crazy_trolley.py
@@ -45,7 +45,7 @@ def run_task(snapshot_config, variant_data, *_):
 
     """
     with LocalTFRunner(snapshot_config=snapshot_config) as runner:
-        n_epochs = 500
+        n_epochs = 1
         n_epoch_cycles = 20
         sampler_batch_size = 5000
         num_timesteps = n_epochs * n_epoch_cycles * sampler_batch_size
@@ -158,7 +158,7 @@ if __name__ == '__main__':
         run_experiment(
             run_task,
             n_parallel=32,
-            snapshot_mode='gap',
+            snapshot_mode='last',
             snapshot_gap=10,
             seed=1,
             plot=False,
