+++ src/ast_toolbox/simulators/crazy_trolley/train_crazy_trolley.py
@@ -45,7 +45,7 @@ def run_task(snapshot_config, variant_data, *_):
 
     """
     with LocalTFRunner(snapshot_config=snapshot_config) as runner:
-        n_epochs = 5
+        n_epochs = 50
         n_epoch_cycles = 20
         sampler_batch_size = 5000
         num_timesteps = n_epochs * n_epoch_cycles * sampler_batch_size
@@ -107,7 +107,7 @@ def run_task(snapshot_config, variant_data, *_):
                    qf=qf,
                    exploration_strategy=epilson_greedy_strategy,
                    replay_buffer=replay_buffer,
-                   qf_lr=1e-4,
+                   qf_lr=1e-2,
                    discount=1.0,
                    min_buffer_size=int(1e4),
                    double_q=True,
@@ -159,7 +159,7 @@ if __name__ == '__main__':
             run_task,
             n_parallel=32,
             snapshot_mode='last',
-            snapshot_gap=10,
+            snapshot_gap=5,
             seed=1,
             plot=False,
             variant={'buffer_size': replay_buffer_size},
