### 2021-06-07 #3312 @mark-koren
> @mark-koren in the paper this dependency should be noted, also in the installation docs. Eventually the best setup to have all software packages "frozen" at the time of the paper, to have a [docker](https://en.wikipedia.org/wiki/Docker_(software)) container description (Dockerfile) in which you can specify fixed version releases for each dependency, including pyhton. It would also facilitate the review actiity.

Sorry, to clarify -- Is this something you need before you can continue the review? Or is this a longer-term add?
