### 2021-06-15 c4c968f
Changes to run trolley diagnostics
