### 2021-07-15 #3312 @sjpfenninger
@mark-koren Can you update on whether you have been able to resolve the package version issues that prevented installation so that @giodegas can continue reviewing? I would be happy with a solution that specifies all dependencies with their versions, it does not have to be full blown Docker solution. Meanwhile I am also chasing up @abhiramm7 who is MIA.
