### 2021-07-26 #3312 @mark-koren
> @mark-koren Can you update on whether you have been able to resolve the package version issues that prevented installation so that @giodegas can continue reviewing? I would be happy with a solution that specifies all dependencies with their versions, **it does not have to be full blown Docker solution.** Meanwhile I am also chasing up @abhiramm7 who is MIA.

I had a docker solution 90% done before I left, but I am digging deeper to find out best practices (e.g. stuff like should I still use venv too, or should docker user be root, etc). I know this level of detail probably isn't needed for the review itself, but I want to make sure the docker option is available to all our users, since this dependency issue seems to be cropping up more.
