+++ requirements.txt
@@ -1,1 +1,196 @@
-.
+absl-py==0.9.0
+aiohttp==3.6.2
+alabaster==0.7.12
+appdirs==1.4.4
+argon2-cffi==20.1.0
+arrow==0.16.0
+astor==0.8.1
+async-generator==1.10
+async-timeout==3.0.1
+atari-py==0.1.15
+attrs==19.3.0
+Babel==2.8.0
+backcall==0.2.0
+bcrypt==3.2.0
+beautifulsoup4==4.9.1
+berkeleydb==18.1.2
+binaryornot==0.4.4
+bleach==3.2.1
+box2d-py==2.3.8
+bresenham==0.2.1
+bsddb3==6.2.7
+build==0.5.1
+bump2version==1.0.0
+bumpversion==0.6.0
+cached-property==1.5.1
+certifi==2020.6.20
+cffi==1.14.3
+chardet==3.0.4
+click==7.1.2
+cloudpickle==1.3
+colorama==0.4.3
+compress-pickle==1.2.0
+configparser2==4.0.0
+cookiecutter==1.7.2
+coverage==5.2.1
+cryptography==3.1.1
+cycler==0.10.0
+dataclasses==0.8
+decorator==4.4.2
+defusedxml==0.7.1
+depq==1.5.5
+dill==0.3.3
+distlib==0.3.1
+docutils==0.16
+entrypoints==0.3
+fabric==2.5.0
+filelock==3.0.12
+fire==0.3.1
+first==2.0.2
+future==0.18.2
+fuzzywuzzy==0.18.0
+garage==2019.10.1
+gast==0.2.2
+gensim==3.8.3
+glfw==1.12.0
+google==3.0.0
+google-pasta==0.2.0
+grpcio==1.30.0
+gym==0.12.4
+h5py==2.10.0
+hickle==4.0.3
+idna==2.10
+idna-ssl==1.1.0
+imageio==2.9.0
+imagesize==1.2.0
+importlib-metadata==1.7.0
+importlib-resources==3.0.0
+iniconfig==1.0.1
+invoke==1.4.1
+ipykernel==5.5.3
+ipython==7.16.1
+ipython-genutils==0.2.0
+ipywidgets==7.6.3
+isort==5.4.2
+jedi==0.18.0
+jeepney==0.4.3
+Jinja2==2.11.2
+jinja2-time==0.2.0
+jsonschema==3.2.0
+jupyter-client==6.1.12
+jupyter-core==4.7.1
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.0
+Keras-Applications==1.0.8
+Keras-Preprocessing==1.1.2
+keyring==21.4.0
+kiwisolver==1.2.0
+Markdown==3.2.2
+MarkupSafe==1.1.1
+matplotlib==3.3.0
+matrix==2.0.1
+mistune==0.8.4
+more-itertools==8.5.0
+msgpack==1.0.0
+multidict==4.7.6
+nbclient==0.5.3
+nbconvert==6.0.7
+nbformat==5.1.3
+nest-asyncio==1.5.1
+networkx==2.4
+notebook==6.3.0
+numpy==1.19.1
+opt-einsum==3.3.0
+packaging==20.4
+pandocfilters==1.4.3
+paramiko==2.7.2
+parso==0.8.2
+Pebble==4.6.0
+pep517==0.10.0
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==7.2.0
+pip-tools==6.1.0
+pkginfo==1.5.0.1
+pluggy==0.13.1
+poyo==0.5.0
+prometheus-client==0.10.1
[…]
