### 2021-08-03 de2c215
Add Docker install instructions to README
