### 2021-08-03 a7a576a
Add Docker for install and running
