+++ docker/entrypoint.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+source venv/bin/activate
+#exec "$@"
+/bin/bash "$@"
