### 2021-08-03 aa43fb8
Change workflow for new docker files
