+++ setup.cfg
@@ -15,7 +15,8 @@ exclude =   .tox,
             dist,
             local,
             scripts,
-            third_party
+            third_party,
+            examples/crazy_trolley
 
 [autoflake]
 exclude = third_party
