### 2021-08-03 fdecfb9
Change workflow for new docker files
