+++ MANIFEST.in
@@ -8,6 +8,7 @@ graft third_party
 graft joss
 recursive-exclude third_party */.git joss
 graft examples
+recursive-exclude examples/crazy_trolley *
 
 include .bumpversion.cfg
 include .coveragerc
@@ -30,6 +31,8 @@ include tox.ini .travis.yml .appveyor.yml
 include test.test
 
 recursive-include docker *.sh
+include .dockerignore
+include docker/Dockerfile
 
 exclude .pre-commit-config.yaml
 exclude .readthedocs.yml
