### 2021-08-03 0173e7c
Bump tensorflow from 1.15.3 to 2.4.2
Bumps [tensorflow](https://github.com/tensorflow/tensorflow) from 1.15.3 to 2.4.2.
- [Release notes](https://github.com/tensorflow/tensorflow/releases)
- [Changelog](https://github.com/tensorflow/tensorflow/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorflow/compare/v1.15.3...v2.4.2)

---
updated-dependencies:
- dependency-name: tensorflow
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
