### 2021-08-03 #3312 @mark-koren
Ok, there is now a docker install available! The instructions are on the README of the github page. Let me know if you run into any issues!
