### 2021-09-05 sisl/AdaptiveStressTestingToolbox#52 @abhiramm7
JOSS Review: Docker Installation 
Hi,

I was trying install the package using docker installation instructions. I am seeing the following error, it seems to be coming from torch. This is on a Mac. Let me know if you are any additional information.

`executor failed running [/bin/sh -c pip install --ignore-installed -r ./requirements.txt]: exit code: 137`
