### 2021-09-06 #3312 @sjpfenninger
@giodegas I realise it's the start of the teaching period, but a quick update is appreciated!
