+++ requirements.txt
@@ -109,7 +109,7 @@ Pebble==4.6.0
 pep517==0.10.0
 pexpect==4.8.0
 pickleshare==0.7.5
-Pillow==7.2.0
+Pillow==9.0.0
 pip-tools==6.1.0
 pkginfo==1.5.0.1
 pluggy==0.13.1
