### 2022-01-21 ac7c419
Bump ipython from 7.16.1 to 7.16.3
Bumps [ipython](https://github.com/ipython/ipython) from 7.16.1 to 7.16.3.
- [Release notes](https://github.com/ipython/ipython/releases)
- [Commits](https://github.com/ipython/ipython/compare/7.16.1...7.16.3)

---
updated-dependencies:
- dependency-name: ipython
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
