+++ requirements.txt
@@ -166,7 +166,7 @@ sphobjinv==2.0.1
 tabulate==0.8.7
 tensorboard==1.15.0
 tensorboardX==2.1
-tensorflow==1.15.3
+tensorflow==2.5.3
 tensorflow-estimator==1.15.1
 tensorflow-probability>=0.8.0
 termcolor==1.1.0
