+++ requirements.txt
@@ -116,7 +116,7 @@ pluggy==0.13.1
 poyo==0.5.0
 prometheus-client==0.10.1
 prompt-toolkit==3.0.18
-protobuf==3.12.2
+protobuf==3.15.0
 psutil==5.7.2
 ptyprocess==0.7.0
 py==1.9.0
