### 2022-04-06 f705fd9
Bump notebook from 6.3.0 to 6.4.10
Bumps [notebook](http://jupyter.org) from 6.3.0 to 6.4.10.

---
updated-dependencies:
- dependency-name: notebook
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
