+++ requirements.txt
@@ -98,7 +98,7 @@ nbconvert==6.0.7
 nbformat==5.1.3
 nest-asyncio==1.5.1
 networkx==2.4
-notebook==6.3.0
+notebook==6.4.10
 numpy==1.19.1
 opt-einsum==3.3.0
 packaging==20.4
