### 2022-04-24 #3312 @arfon
:wave: all, I'm taking over this review as @sjpfenninger is stepping down from the JOSS editorial team. Thanks for all of your efforts @sjpfenninger!

@giodegas – my understanding was that you were waiting to start your review in earnest until there was a Docker solution in place which it looks like there now is (https://github.com/openjournals/joss-reviews/issues/3312#issuecomment-892157745).
