### 2022-05-03 #3312 @mossr
Hey @mark-koren, it may make sense to cite my related JOSS paper (https://joss.theoj.org/papers/10.21105/joss.02749).
```
@article{moss2021pomdpstresstesting,
  title = {{POMDPStressTesting.jl}: Adaptive Stress Testing for Black-Box Systems},
  author = {Robert J. Moss},
  journal = {Journal of Open Source Software},
  year = {2021},
  volume = {6},
  number = {60},
  pages = {2749},
  doi = {10.21105/joss.02749}
}
```
