+++ requirements.txt
@@ -31,7 +31,7 @@ cloudpickle==1.3
 colorama==0.4.3
 compress-pickle==1.2.0
 configparser2==4.0.0
-cookiecutter==1.7.2
+cookiecutter==2.1.1
 coverage==5.2.1
 cryptography==3.1.1
 cycler==0.10.0
