+++ requirements.txt
@@ -99,7 +99,7 @@ nbformat==5.1.3
 nest-asyncio==1.5.1
 networkx==2.4
 notebook==6.3.0
-numpy==1.19.1
+numpy==1.22.0
 opt-einsum==3.3.0
 packaging==20.4
 pandocfilters==1.4.3
