+++ requirements.txt
@@ -89,7 +89,7 @@ Markdown==3.2.2
 MarkupSafe==1.1.1
 matplotlib==3.3.0
 matrix==2.0.1
-mistune==0.8.4
+mistune==2.0.3
 more-itertools==8.5.0
 msgpack==1.0.0
 multidict==4.7.6
