### 2022-08-10 43a123d
Bump nbconvert from 6.0.7 to 6.3.0
Bumps [nbconvert](https://github.com/jupyter/nbconvert) from 6.0.7 to 6.3.0.
- [Release notes](https://github.com/jupyter/nbconvert/releases)
- [Commits](https://github.com/jupyter/nbconvert/compare/6.0.7...6.3.0)

---
updated-dependencies:
- dependency-name: nbconvert
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
