### 2022-09-17 #3312 @arfon
:wave: @mark-koren – I just heard back from @giodegas that they won't be able to complete their review. I will provide second review myself. Before I do so however I would like you to confirm that you're still interested in pursing this publication with JOSS?

I ask as it looks like you didn't respond to https://github.com/sisl/AdaptiveStressTestingToolbox/issues/53 or https://github.com/sisl/AdaptiveStressTestingToolbox/issues/52 from @abhiramm7's review.
